+++ test/test_group_tableaux.jl
@@ -1,11 +1,10 @@
-@testitem "Classical" begin
+@testitem "group theory routines" begin
     using Test
-
     using Random
     using QuantumClifford
 
     # Including sizes that would test off-by-one errors in the bit encoding.
-    test_sizes = [1, 2, 3, 4, 5, 7, 8, 9, 15, 16, 17] 
+    test_sizes = [1, 2, 3, 4, 5, 7, 8, 9, 15, 16, 17]
     # Zero function(in groupify) slows down around 2^30(n=30),eventually breaks
     small_test_sizes = [1, 2, 3, 4, 5, 7] # Pauligroup slows around n = 8
 
@@ -45,7 +44,7 @@
             end
         end
         #Test pauligroup
-        for n in [1, small_test_sizes...] 
+        for n in [1, small_test_sizes...]
             @test length(QuantumClifford.pauligroup(n, phases=false)) == 4^n
             @test length(QuantumClifford.pauligroup(n, phases=true)) == 4^(n+1)
         end
@@ -98,7 +97,7 @@
             end
             c = contractor(s, subset)
             count = 0
-            for stabilizer in s 
+            for stabilizer in s
                 contractable = true
                 for i in subset
                     if stabilizer[i] != (false, false) contractable = false end
@@ -111,9 +110,9 @@
                     p = zero(PauliOperator, nqubits(s))
                     index = 0
                     for i in 1:nqubits(s)
-                        if !(i in subset) 
-                            index+=1 
-                            p[i] = contracted[index] 
+                        if !(i in subset)
+                            index+=1
+                            p[i] = contracted[index]
                         end
                     end
                     @test p in s || -1* p in s || 1im * p in s || -1im * p in s