+++ test/test_projections.jl
@@ -76,12 +76,12 @@
         @test_throws BadDataStructure pds, a, r = project!(copy(ds),p)
         pms, a, r = project!(copy(ms),p)
         @test mixed_stab_looks_good(pms)
-        @test pms.rank==3
-        @test a==pms.rank && isnothing(r)
+        @test rank(pms)==3
+        @test a==rank(pms) && isnothing(r)
         pmds, a, r = project!(copy(mds),p)
         @test mixed_destab_looks_good(pmds)
-        @test pmds.rank==3
-        @test a==pmds.rank && isnothing(r)
+        @test rank(pmds)==3
+        @test a==rank(pmds) && isnothing(r)
 
         p = P"ZZI"
         ps, a, r = project!(copy(s),p)
@@ -90,11 +90,11 @@
         @test_throws BadDataStructure pds, a, r = project!(copy(ds),p)
         pms, a, r = project!(copy(ms),p)
         @test mixed_stab_looks_good(pms)
-        @test pms.rank==2
+        @test rank(pms)==2
         @test a==0 && r==0x2
         pmds, a, r = project!(copy(mds),p)
         @test mixed_destab_looks_good(pmds)
-        @test pmds.rank==2
+        @test rank(pmds)==2
         @test a==0 && r==0x2
         @test canonicalize!(ps)==canonicalize!(stabilizerview(pms))==canonicalize!(stabilizerview(pmds))
 
@@ -107,11 +107,11 @@
         @test a==2 && isnothing(r)
         pms, a, r = project!(copy(ms),p)
         @test mixed_stab_looks_good(pms)
-        @test pms.rank==2
+        @test rank(pms)==2
         @test a==2 && isnothing(r)
         pmds, a, r = project!(copy(mds),p)
         @test mixed_destab_looks_good(pmds)
-        @test pmds.rank==2
+        @test rank(pmds)==2
         @test a==2 && isnothing(r)
         @test canonicalize!(ps)==canonicalize!(stabilizerview(pms))==canonicalize!(stabilizerview(pds))==canonicalize!(stabilizerview(pmds))
     end
@@ -162,10 +162,10 @@
         s = MixedStabilizer(s, 2)
         ms, a, r = project!(copy(s), P"IZI")
         @test (a, r) == (0, 0x0) # on commuting operator in the stabilizer
-        @test ms.rank == 2
+        @test rank(ms) == 2
         ms, a, r = project!(copy(s), P"IIZ")
         @test (a, r) == (3, nothing) # on commuting operator out of the stabilizer
-        @test ms.rank == 3
+        @test rank(ms) == 3
         s = S"ZII IZI"
         s = Destabilizer(s)
         @test_throws BadDataStructure project!(copy(s), P"IZI"; keep_result=true)  # on comm
@@ -182,16 +182,16 @@
         s = MixedDestabilizer(s)
         mds, a, r = project!(copy(s), P"IZI"; keep_result=true)
         @test (a, r) == (0, 0x0) # on commuting operator in the stabilizer
-        @test mds.rank == 2
+        @test rank(mds) == 2
         mds, a, r = project!(copy(s), P"IIZ"; keep_result=true)
         @test (a, r) == (3, nothing) # on commuting operator out of the stabilizer
-        @test mds.rank == 3
+        @test rank(mds) == 3
         mds, a, r = project!(copy(s), P"IZI"; keep_result=false)
         @test (a, r) == (0, nothing) # on commuting operator in the stabilizer
-        @test mds.rank == 2
+        @test rank(mds) == 2
         mds, a, r = project!(copy(s), P"IIZ"; keep_result=false)
         @test (a, r) == (3, nothing) # on commuting operator out of the stabilizer
-        @test mds.rank == 3
+        @test rank(mds) == 3
     end
     @testset "Results from canonicalization vs from destabilizer" begin
         @test generate!(P"_Z", S"XZ") === nothing # for bug fixed in 4b536231c3ee4e6446262fcc61ba8da669415bc8
@@ -206,7 +206,7 @@
                 _, ams, rms = project!(ms,p)
                 _, amd, rmd = project!(md,p)
                 @test rs == rms == rmd
-                @test (md.rank!=r) || (canonicalize!(s) == canonicalize!(stabilizerview(ms)))
+                @test (rank(md)!=r) || (canonicalize!(s) == canonicalize!(stabilizerview(ms)))
                 @test canonicalize!(stabilizerview(ms)) == canonicalize!(stabilizerview(md))
                 if as == 0
                     @test ams == amd == 0